+++ modules/server/src/main/ts/bedrock/compiler/Webpack.ts
@@ -153,6 +153,9 @@ const getWebPackConfigTs = (tsConfigFile: string, scratchFile: string, dest: str
       // See https://github.com/TypeStrong/ts-loader#usage-with-webpack-watch
       new webpack.WatchIgnorePlugin({
         paths: [ /\.d\.ts$/ ]
+      }),
+      new webpack.DefinePlugin({
+        'process.env.NODE_ENV': JSON.stringify('development')
       })
     ],
 
@@ -204,6 +207,11 @@ const getWebPackConfigJs = (scratchFile: string, dest: string, coverage: string[
         ] : []
       )
     },
+    plugins: [
+      new webpack.DefinePlugin({
+        'process.env.NODE_ENV': JSON.stringify('development')
+      })
+    ],
     output: {
       filename: path.basename(dest),
       path: path.resolve(path.dirname(dest))